eos name and symbol parsing
